move show impls to own file
